Lots of general refactorings and improvements for maintainablity around DAO objects
