+++ src/figgy/data/models/config_item.py
@@ -1,10 +1,7 @@
 import time
+
 from enum import Enum
-from typing import Dict
-
-from pydantic import dataclasses
-
-from figgy.constants.data import *
+from pydantic import Field, validator, BaseModel
 
 
 class ConfigState(Enum):
@@ -12,19 +9,17 @@ class ConfigState(Enum):
     ACTIVE = 1
 
 
-@dataclasses.dataclass(frozen=True)
-class ConfigItem:
-    name: str
+class ConfigItem(BaseModel):
+    name: str = Field(None, alias='parameter_name')
     state: ConfigState
-    last_updated: int
+    last_updated: int = time.time()
 
-    @staticmethod
-    def from_dict(obj: Dict) -> "ConfigItem":
-        name = obj.get(CACHE_PARAMETER_KEY_NAME, None)
-        last_updated = obj.get(CACHE_LAST_UPDATED_KEY_NAME, time.time())
-        state = obj.get(CACHE_STATE_ATTR_NAME, None)
-
-        return ConfigItem(name=name, last_updated=last_updated, state=ConfigState[state])
+    @validator('state', pre=True)
+    def init_state(cls, value):
+        return ConfigState[value]
 
     def __lt__(self, o: "ConfigItem") -> bool:
         return self.last_updated < o.last_updated
+
+    def __hash__(self):
+        return hash(f'{self.name}-{self.state.name}-{self.last_updated}')
