+++ src/figgy/models/string_serializable.py
@@ -1,0 +1,11 @@
+from pydantic.decorator import BaseModel
+
+
+class StringSerializable(BaseModel):
+    """
+        Classes extending (implementing) StringSerializable should implement serialization to a single string
+        by overriding the dict() and json() methods.
+    """
+
+    def dict(self, **kwargs):
+        pass
